Fix issue with evil-cleverparens text objects
When using the inside variants of text objects, evil-cleverparens for
some reason decides to not move the cursor inside the text object if
the cursor was not positioned inside the text object to begin
with. Evil itself handles this correctly but evil-cleverparens
specifically opts out of it. This change just removes the line that
opts out of that behaviour.

https://github.com/luxbock/evil-cleverparens/pull/71